requirements.txt
numpy>=1.22.1
scipy>=1.7.3
matplotlib>=3.5.1
scikit-learn==1.2.0
pyriemann>=0.2.7
mne>=1.0.2
joblib>=1.1.0
pandas>=1.4.0
setuptools>=60.7.0

# --------- dask --------- #
dask
dask-ml

h5py

# --------- optuna --------- #
optuna>=2.10.0
# torch==2.0.1+cu117
# --------- pytorch --------- #
lightning>=2.0.0
torchsummary
einops

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
# wandb
# neptune-client
mlflow
# aim-mlflow
# comet-ml
# aim>=3.16.2  # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550

# --------- others --------- #
pyrootutils     # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
pytest-cov
docstr-coverage
# sh            # for running bash commands in some tests (linux/macos only)